# openapi.yaml
swagger: "2.0"
info:
  description: "TODO BA-This is the api reference for the Block Array dApp which allows recording, transfering and signing of bills of lading."
  version: "1.0.0"
  title: "TODO BA- Block Array Bill of Ladding dApp"
  termsOfService: "TODO BA"
  contact:
    email: "TODO@ba.com"
  license:
    name: "TODO BA"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html" #TODO BA
host: "petstore.swagger.io" #TODO BA
basePath: "/"
...
schemes:
- "https"
paths:
...
  /proxy/getTarget/:
    get:
      tags:
      - "Proxy and Upgrade"
      summary: "Gets address of Function"
      description: "Gets address that is delegate called to run a function's logic."
      operationId: "getTarget"
      parameters:
      - name: "id"
        in: "query"
        description: "The function selector the return the target of"
        required: true
        type: "string"
        maxLength: 10
        minLength: 10
      produces:
      - "application/xml"
      - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
        425:
          description: "Invalid Input"
...
  /proxy/setTarget/:
    post:
      tags:
      - "Proxy and Upgrade"
      summary: "Sets the address of Function"
      description: "Sets address that is delegate called to run a function's logic. This function must be called with an signed nonce from the Block Array account in the https body."
      operationId: "setTarget"
      parameters:
      - name: "id"
        in: "query"
        description: "Function selector whose target to change."
        required: true
        type: "string"
        minLength: 10
        maxLength: 10
      - name: "target"
        in: "query"
        description: "The address to set the function too"
        required: true
        type: "string"
        minLength: 42
        maxLength: 42
      - in: "body"
        name: "auth"
        description: "The Block Array signature of the nonce"
        required: true
        schema:
          $ref: "#/definitions/auth"
      produces:
      - "application/xml"
      - "application/json"
      responses:
        200:
          description: "Successfuly Changed"
          schema:
            $ref: "#/definitions/ApiResponse"
        425:
          description: "Invalid Input"
  /proxy/changeMaster/:
    post:
      tags:
      - "Proxy and Upgrade"
      summary: "Sets the address of Function"
      description: "Sets address that is delegate called to run a function's logic. This function must be called with an signed nonce from the Block Array account in the https body."
      operationId: "changeMaster"
      parameters:
      - name: "newMaster"
        in: "query"
        description: "The new master address"
        required: true
        type: "string"
        minLength: 42
        maxLength: 42
      - in: "body"
        name: "auth"
        description: "The Block Array signature of the nonce"
        required: true
        schema:
          $ref: "#/definitions/auth"
      produces:
      - "application/xml"
      - "application/json"
      responses:
        200:
          description: "Successfuly Changed"
          schema:
            $ref: "#/definitions/ApiResponse"
        425:
          description: "Invalid Input"
  /proxy/changeController/:
    post:
      tags:
      - "Proxy and Upgrade"
      summary: "Sets the the new controller address"
      description: "Sets the new controller address."
      operationId: "changeController"
      parameters:
      - name: "newController"
        in: "query"
        description: "The new controller address"
        required: true
        type: "string"
        minLength: 42
        maxLength: 42
      - in: "body"
        name: "auth"
        description: "The Block Array signature of the nonce"
        required: true
        schema:
          $ref: "#/definitions/auth"
      produces:
      - "application/xml"
      - "application/json"
      responses:
        200:
          description: "Successfuly Changed"
          schema:
            $ref: "#/definitions/ApiResponse"
        425:
          description: "Invalid Input"
  /proxy/pause:
    post:
      tags:
      - "Proxy and Upgrade"
      summary: "Pauses the execution"
      description: "Calls the pause method in the universal proxy, blocking any state changes while Blockarray updates."
      operationId: "pause"
      parameters:
      - in: "body"
        name: "auth"
        description: "The Block Array signature of the nonce"
        required: true
        schema:
          $ref: "#/definitions/auth"
      produces:
      - "application/xml"
      - "application/json"
      responses:
        200:
          description: "Successfuly Paused"
          schema:
            $ref: "#/definitions/ApiResponse"
        425:
          description: "Invalid Input"
definitions:
  auth:
    type: "object"
    required:
    - "signature"
    properties:
      signature:
        type: "string"
        minLength: 132 
        maxLength: 132 
  ApiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      type:
        type: "string"
      message:
        type: "string"
...